# perf(startup): use WAL journal for sqlite databases in DENO_DIR